Update IntelliJ plugin SDKs to 3.6 Canary 5 (3.6.0.5)
The SDK unfortunately needs a manual modification to work around b/?.
The manual change was to remove the file
ProductivityFeaturesRegistry.xml from inside lib/idea.jar as it
conflicts with the file of the same name present in resources.jar.
Having both of them results in test failures such as
//third_party/intellij/bazel/plugin/base:integration_tests failures in sponge/b023dbaf-b311-46ac-a9bc-9ce61440fccf

This CL also upstreams necessary changes to BlazeModuleSystem and
BlazeProjectSystemSyncManager, both necessary to keep up with the API
changes in Canary 5.

BlazeModuleSystem also includes a new API that depends on a newly
introduced enum ScopeType. This enum is re-defined in the sdkcompat
for 3.4 and 3.5 in order to prevent compile failures.

PiperOrigin-RevId: 259735604