Set android SDK from project view if one doesn't exist after sync.
Currently, a directory only sync will not update the project's SDK.
In a large android project, this means the android SDK won't be set
until the next "real" sync succeeds. The SDK is required for many
IDE functionalities to work, such as AdbService and attaching to
devices.

Normally info regarding a project's SDK is retained between IDE
restarts.  However when the user imports a new project, they face
a special situation where project SDK is not available AND the init
sync usually takes a very long time. When this happens, the user won't
be able to use IDE functionalities that require an android SDK.

This CL changes BlazeAndroidSyncPlugin#updateProjectSdk to set
the project's SDK from project view if:
1. There is no SDK set.
2. There is no android sync data from the sync result.

This way the user will have a usable SDK the moment after the
initial directory only sync finishes, instead of having to wait for
a real sync to complete.

PiperOrigin-RevId: 265685251